Add files via upload
Hints for below questions:
Question 2
A simple linear classifier can represent which of the following functions? (Check all that apply)
Hint: If you are stuck, see 
https://www.coursera.org/learn/ml-foundations/module/nqC1t/discussions/AAIUurrtEeWGphLhfbPAyQ
Question 3
Which of the the following neural networks can represent the following function? Select all that apply.
(x1 AND x2) OR (NOT x1 AND NOT x2)